break out component and component factory into new files.
